repo: move graph-related staticmethods into graph.py
As per @Suor's request.
